back to floats

+++ cpp/predictor.h
@@ -19,14 +19,14 @@ class Predictor {
 public:
     Predictor(int num_users, int num_movies, int nf);
     ~Predictor();
-    double predict(int user, short movie);
+    float predict(int user, short movie);
     int get_num_users() { return num_users; }
     int get_num_movies() { return num_movies; }
     int get_num_features() { return num_features; }
 
     // the real meat of the predictor
-    double *movie_features;
-    double *user_features;
+    float *movie_features;
+    float *user_features;
 
 private:
     int num_users;
